Finish the main build
I'm really happy with how this turned out. It looks pretty good
on the screen and works really well with almost everything that I
want it to be able to do.

I am still yet to make it so that you can input decimals but this
shouldn't be too difficult.

I also need to make it so that numbers too big for the screen don't go
beyond the edge of the screen but this shouldn't be too hard.